move GatherAndCompare to tesutil
